feat(frontend): conditionally reset UTXOS store on BTC conversion
