+++ src/frontend/src/eth/components/fee/FeeContext.svelte
@@ -154,9 +154,9 @@
 	$: obverseFeeData(observe);
 
 	$: $ckEthMinterInfoStore,
-		() => {
+		(() => {
 			observe && debounceUpdateFeeData();
-		};
+		})();
 
 	/**
 	 * Expose a call to evaluate, so that consumers can re-evaluate imperatively, for example, when the amount or destination is manually updated by the user.
